+++ bindings/nostr-sdk-ffi/src/client/mod.rs
@@ -29,7 +29,7 @@ use self::zapper::{ZapDetails, ZapEntity};
 use crate::abortable::AbortHandle;
 use crate::error::Result;
 use crate::relay::options::{NegentropyOptions, SubscribeAutoCloseOptions};
-use crate::relay::RelayOptions;
+use crate::relay::{RelayBlacklist, RelayOptions};
 use crate::{HandleNotification, NostrDatabase, Relay};
 
 #[derive(Object)]
@@ -85,6 +85,59 @@ impl Client {
 
     pub fn database(&self) -> Arc<NostrDatabase> {
         Arc::new(self.inner.database().into())
+    }
+
+    /// Get blacklist
+    pub fn blacklist(&self) -> RelayBlacklist {
+        self.inner.blacklist().into()
+    }
+
+    /// Mute event IDs
+    ///
+    /// Add event IDs to blacklist
+    /// 
+    /// <div class="warning">Mute list event is not currently created/updated!</div>
+    pub fn mute_ids(&self, ids: Vec<Arc<EventId>>) {
+        block_on(async { self.inner.mute_ids(ids.into_iter().map(|id| **id)).await })
+    }
+
+    /// Unmute event IDs
+    ///
+    /// Remove event IDs from blacklist
+    /// 
+    /// <div class="warning">Mute list event is not currently created/updated!</div>
+    pub fn unmute_ids(&self, ids: &[Arc<EventId>]) {
+        block_on(async {
+            self.inner
+                .unmute_ids(ids.iter().map(|id| id.as_ref().deref()))
+                .await
+        })
+    }
+
+    /// Mute public keys
+    ///
+    /// Add public keys to blacklist
+    /// 
+    /// <div class="warning">Mute list event is not currently created/updated!</div>
+    pub fn mute_public_keys(&self, public_keys: Vec<Arc<PublicKey>>) {
+        block_on(async {
+            self.inner
+                .mute_public_keys(public_keys.into_iter().map(|p| **p))
+                .await
+        })
+    }
+
+    /// Unmute public keys
+    ///
+    /// Remove public keys from blacklist
+    /// 
+    /// <div class="warning">Mute list event is not currently created/updated!</div>
+    pub fn unmute_public_keys(&self, public_keys: &[Arc<PublicKey>]) {
+        block_on(async {
+            self.inner
+                .unmute_public_keys(public_keys.iter().map(|p| p.as_ref().deref()))
+                .await
+        })
     }
 
     pub fn start(&self) {