+++ bindings/nostr-sdk-js/src/relay/mod.rs
@@ -13,6 +13,7 @@ use nostr_js::types::JsFilter;
 use nostr_sdk::prelude::*;
 use wasm_bindgen::prelude::*;
 
+pub mod blacklist;
 pub mod flags;
 pub mod limits;
 pub mod options;