+++ crates/nostr-relay-pool/src/lib.rs
@@ -23,4 +23,4 @@ pub use self::relay::options::{
     SubscribeAutoCloseOptions, SubscribeOptions,
 };
 pub use self::relay::stats::RelayConnectionStats;
-pub use self::relay::{Relay, RelayNotification, RelayStatus};
+pub use self::relay::{Relay, RelayBlacklist, RelayNotification, RelayStatus};